requirements.txt
flask
flask_cors
numpy
pandas
requests
matplotlib
cryptography
mysql-connector-python
